--- src/main/java/Command.java
@@ -1,5 +1,0 @@
-abstract public class Command {
-    protected boolean isExit = false;
-
-    public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws GabbyException;
-}
