+++ src/main/java/gabby/GabbyException.java
@@ -1,3 +1,5 @@
+package gabby;
+
 public class GabbyException extends Exception {
     public GabbyException(String errorMsg) {
         super(errorMsg);
