+++ src/main/java/gabby/command/ByeCommand.java
@@ -1,3 +1,9 @@
+package gabby.command;
+
+import gabby.Storage;
+import gabby.Ui;
+import gabby.task.TaskList;
+
 public class ByeCommand extends Command {
     public ByeCommand() {
         super.isExit = true;
