+++ src/main/java/gabby/command/Command.java
@@ -1,0 +1,16 @@
+package gabby.command;
+
+import gabby.GabbyException;
+import gabby.Storage;
+import gabby.Ui;
+import gabby.task.TaskList;
+
+abstract public class Command {
+    protected boolean isExit = false;
+
+    public boolean isExit() {
+        return this.isExit;
+    }
+
+    public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws GabbyException;
+}
