+++ src/main/java/gabby/task/EventTask.java
@@ -1,3 +1,7 @@
+package gabby.task;
+
+import gabby.GabbyException;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeParseException;
